Refactor integration_tests
