+++ features/solvers.feature
@@ -1,10 +1,10 @@
-Feature: Check for solvers available in Thoth
+Feature: Check for minimum solvers available in Thoth
     Scenario: Check there are minimum solvers available in Thoth
-        Given a set of solvers
+        Given deployment is accessible using HTTPS
+        And a minimum set of solvers requested
             | solver_name |
             | solver-rhel-8-py36 |
             | solver-fedora-31-py37 |
             | solver-fedora-31-py38 |
-        And deployment is accessible using HTTPS
         When we ask for the available solvers
-        Then they should include at least the given solvers
+        Then they should include at least the minimum set of solvers
