+++ features/steps/index.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-"""Basic integration tests for Thoth deployment."""
+"""Thoth indices integration test for Thoth deployment."""
 
 import os
 import requests
@@ -25,42 +25,18 @@ from behave import given, when, then
 from hamcrest import assert_that, equal_to
 
 
-@given('I query Thoth User API for "{index}" "{package}" "{version}"')
-def step_impl(context, index: str, package: str, version: str):
-    """Retrieve metadata about Python Package."""
-    context.query_string = f"{context.scheme}://{context.user_api_url}/api/v1/"
-    f"python/package/metadata?index={urllib.parse.quote_plus(index)}&name={package}&version={version}"
-
-
-@when("I query for the list of known Python Package indices,")
+@when("I query for the list of known Python Package indices")
 def step_impl(context):
     """Retrieve list of Python Package Indices known to Thoth."""
+    context.result = {}
     response = requests.get(f"{context.scheme}://{context.user_api_url}/api/v1/python-package-index", verify=False,)
     assert (
         response.status_code == 200
     ), f"Bad status code ({response.status_code}) when obtaining python-package-index from {context.user_api_url}"
-    context.result = response.json()
-
-
-@when("I execute the query")
-def step_impl(context):
-    """Execute the prepared query."""
-    response = requests.get(f"{context.query_string}", verify=False,)
-    assert (
-        response.status_code == 200
-    ), f"Bad status code ({response.status_code}) when executing the query: {context.query_string}"
-
-    context.result = response.json()
-
-
-@then('I should get "{author}" and "{maintainer}"')
-def step_impl(context, author: str, maintainer: str):
-    """Verify conditions for author and maintainer."""
-    assert_that(context.result["author"], equal_to(author))
-    assert_that(context.result["maintainer"], equal_to(maintainer))
+    context.result["indices"] = response.json()
 
 
 @then('I should get a list of "{number}"')
 def step_impl(context, number: int):
     """Verify correct number is retrieved."""
-    assert_that(len(context.result), equal_to(int(number)))
+    assert_that(len(context.result["indices"]), equal_to(int(number)))
